25-08-2011

+++ lib/classes/polling.class.php
@@ -122,7 +122,6 @@ class Default_Model_Polling_Class
 						$nameArr = pathinfo($v1['file']);
 						$nameArr['reversename'] = strrev( $nameArr['filename'] );
 						$nameArr['shortname'] = strrev(substr(strstr($nameArr['reversename'],'-'),1));
-
 						$result0 = $this->m_mysqli->query("
 							SELECT DISTINCT wf_index, wf_count, wf_status, wf_filename 
 							FROM watch_file as wf, queue_commands as cq 
@@ -130,8 +129,6 @@ class Default_Model_Polling_Class
 						if ($result0->num_rows ==0) {
 								$result = $this->m_mysqli->query("	INSERT INTO `watch_file` (`wf_filename`, `wf_fileoutname`, `wf_folder`, `wf_extension`, `wf_filesize0`, `wf_filedate0`, `wf_count`, `wf_status`) 
 																	VALUES ( '".$nameArr['shortname']."',  '".$v1['file']."', '".$k0."', '.".$nameArr['extension']."', '".$v1['size']."', '".$v1['date']."', '0', 'W' )");
-//								echo "INSERT INTO `watch_file` (`wf_filename`, `wf_fileoutname`, `wf_folder`, `wf_extension`, `wf_filesize0`, `wf_filedate0`, `wf_count`, `wf_status`) 
-//																	VALUES ( '".$nameArr['shortname']."',  '".$v1['file']."', '".$k0."', '.".$nameArr['extension']."', '".$v1['size']."', '".$v1['date']."', '0', 'W' )";
 							$wf_row = $this->m_mysqli->insert_id;
 						}else{
 							$row0 = $result0->fetch_object();
@@ -143,13 +140,13 @@ class Default_Model_Polling_Class
 							$wf_row = $row0->wf_index;
 						}
 
-// Check this any of this set of files is still on the queue if not unwatch and unlink
+// Check if any of this set of files in the read output tree array are still on the command_queue or older than 24 hours. If so unlink and unwatch 
 
 						$result5 = $this->m_mysqli->query("
 							SELECT * 
 							FROM queue_commands 
 							WHERE cq_filename LIKE '".$nameArr['shortname']."%'");
-						if ($result5->num_rows == 0) {
+						if ($result5->num_rows == 0 && (time() > $v1['date'] + 86399)) {						
 							$unlink_file= $destination['encoder-output'].$k0."/".$v1['file'];
 							unlink($unlink_file);
 							$this->m_mysqli->query("
@@ -169,23 +166,29 @@ class Default_Model_Polling_Class
 			FROM watch_file as wf, queue_commands as cq, workflow_map as wm 
 			WHERE cq.cq_filename=wf.wf_filename AND wm.wm_workflow=wf.wf_folder AND wm.wm_mimetype=wf.wf_extension AND cq.cq_step = '4' AND wf.wf_filesize0=wf.wf_filesize1 AND wf.wf_filesize1=wf.wf_filesize2 AND cq.cq_status IN ('N','D') AND wf.wf_count!=0 AND wf.wf_status='W' AND wm.wm_media_root='./feeds' ");
 
+// Used with the new workflow_map1 
+/*	$result3 = $this->m_mysqli->query("
+			SELECT * 
+			FROM watch_file as wf, queue_commands as cq, workflow_map1 as wm 
+			WHERE cq.cq_filename=wf.wf_filename AND wm.wm_mimetype=wf.wf_extension AND cq.cq_step = '4' AND wf.wf_filesize0=wf.wf_filesize1 AND wf.wf_filesize1=wf.wf_filesize2 AND cq.cq_status IN ('N','D') AND wf.wf_count!=0 AND wf.wf_status='W' ");
+*/
 		while ($row3 = $result3->fetch_object()) {
 			$result4 = $this->m_mysqli->query("
 				UPDATE queue_commands 
 				SET `cq_update` = '".date("Y-m-d H:i:s", time())."' ,`cq_status`= 'D', cq_result='".$row3->cq_data."' 
 				WHERE cq_index='".$row3->cq_index."' ");
-			$mData = json_decode($row3->cq_data, true);
+			$mData = unserialize($row3->cq_data);
 			$mData['source_filename'] = $row3->wf_fileoutname;
 			$mData['destination_filename'] = $row3->wf_filename.$row3->wm_target_filename;
 			if ($row3->wm_target_folder!='' || $row3->wm_target_folder!='/')
 				$mData['destination_path'] .= $row3->wm_target_folder;
-			$cqData=json_decode($row3->cq_data, true);
+			$cqData = unserialize($row3->cq_data, true);
 			$mData['original_filename'] = $cqData['source_filename'];
 			$mData['flavour'] = $row3->wm_flavour;
 			$mData['duration'] = '1.35';
 			$this->m_mysqli->query("
 				INSERT INTO queue_commands (`cq_command`, `cq_filename`, `cq_cq_index`, `cq_mq_index`, `cq_step`, `cq_data`, `cq_result`, `cq_time`, `cq_update`, `cq_status`) 
-				VALUES ('".$row3->cq_command."', '".$row3->wf_fileoutname."', '".$row3->cq_cq_index."', '".$row3->cq_mq_index."', '".$row3->cq_step."', '".$row3->cq_data."',  '".json_encode($mData)."', '".$row3->cq_time."', '".date("Y-m-d H:i:s", time())."', 'Y')");
+				VALUES ('".$row3->cq_command."', '".$row3->wf_fileoutname."', '".$row3->cq_cq_index."', '".$row3->cq_mq_index."', '".$row3->cq_step."', '".$row3->cq_data."',  '".serialize($mData)."', '".$row3->cq_time."', '".date("Y-m-d H:i:s", time())."', 'Y')");
 
 //	print_r($mData);
 			$result4 = $this->m_mysqli->query("
